+++ packages/components-angular/tests/unit/config/vitest.config.ts
@@ -1,13 +1,6 @@
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
-  esbuild: {
-    tsconfigRaw: {
-      compilerOptions: {
-        experimentalDecorators: true,
-      },
-    },
-  },
   test: {
     globals: false,
     environment: 'jsdom',