+++ packages/components-react/projects/nextjs/components/Toast.tsx
@@ -3,7 +3,7 @@
 import { PToast, useToastManager } from '@porsche-design-system/components-react/ssr';
 import { useEffect } from 'react';
 
-export const Toast = ({ text }: { text: string }): JSX.Element => {
+export const Toast = ({ text }: { text: string }) => {
   const { addMessage } = useToastManager();
   useEffect(() => {
     addMessage({ text });