+++ packages/components-react/src/components/Toast.tsx
@@ -1,7 +1,7 @@
 import { PToast, useToastManager } from '@porsche-design-system/components-react';
 import { useEffect } from 'react';
 
-export const Toast = ({ text }: { text: string }): JSX.Element => {
+export const Toast = ({ text }: { text: string }) => {
   const { addMessage } = useToastManager();
   useEffect(() => {
     addMessage({ text });