+++ packages/components/scripts/wrapper-generator/InputParser.ts
@@ -1,7 +1,7 @@
 import { isDeprecatedComponent } from '@porsche-design-system/component-meta/src/utils';
 import type { TagName } from '@porsche-design-system/shared';
 import * as fs from 'fs';
-import { globbySync } from 'globby';
+import { sync as globbySync } from 'fast-glob';
 import * as path from 'path';
 
 const ROOT_DIR = path.normalize(__dirname + '/../../');