+++ packages/components/CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 - `Link Pure`: focus border of slotted anchor 
    ([#4285](https://github.com/porsche-design-system/porsche-design-system/pull/4285))
+- `Tailwind CSS`: Custom tailwind prefixing
+  ([#4294](https://github.com/porsche-design-system/porsche-design-system/pull/4294))
 
 ## [4.0.0-rc.1] - 2026-04-07